Rename common/data/rxmodel to workbook and move MimeType.
