--- src/main/kotlin/org/wycliffeassociates/otter/common/data/rxmodel/TextItem.kt
@@ -1,6 +1,0 @@
-package org.wycliffeassociates.otter.common.data.rxmodel
-
-data class TextItem(
-    val text: String,
-    val format: MimeType
-)
