+++ src/main/kotlin/org/wycliffeassociates/otter/common/data/workbook/Take.kt
@@ -1,6 +1,7 @@
-package org.wycliffeassociates.otter.common.data.rxmodel
+package org.wycliffeassociates.otter.common.data.workbook
 
 import com.jakewharton.rxrelay2.BehaviorRelay
+import org.wycliffeassociates.otter.common.data.model.MimeType
 import java.io.File
 import java.time.LocalDate
