+++ src/main/kotlin/org/wycliffeassociates/otter/common/data/workbook/TextItem.kt
@@ -1,0 +1,8 @@
+package org.wycliffeassociates.otter.common.data.workbook
+
+import org.wycliffeassociates.otter.common.data.model.MimeType
+
+data class TextItem(
+    val text: String,
+    val format: MimeType
+)
