+++ src/main/kotlin/org/wycliffeassociates/otter/common/data/workbook/Workbook.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.data.rxmodel
+package org.wycliffeassociates.otter.common.data.workbook
 
 data class Workbook(
     val source: Book,
